moved files out of lib/mufon_news to lib - updated gemfile info
