+++ Gemfile
@@ -1,6 +1,8 @@
 source "https://rubygems.org"
 
-git_source(:github) {|repo_name| "https://github.com/#{repo_name}" }
+git_source(:github) {|repo_name| "https://github.com/#{TrevorSeitz/mufon_news}" }
 
 # Specify your gem's dependencies in mufon_news.gemspec
 gemspec
+require "nokogiri"
+require "open-uri"
