+++ bin/console
@@ -1,7 +1,7 @@
 #!/usr/bin/env ruby
 
 require "bundler/setup"
-require "mufon_news"
+# require "mufon_news"
 
 # You can add fixtures and/or initialization code here to make experimenting
 # with your gem easier. You can also use a different console, if you like.
@@ -11,7 +11,7 @@ require "mufon_news"
 # Pry.start
 
 require "nokogiri"
+# require "openuri"
 require "irb"
 # IRB.start(__FILE__)
 
-CommandLineInterface.new
