+++ bin/run_mufon_scraper
@@ -1,6 +1,6 @@
 #!/usr/bin/env ruby
 
-require_relative "../lib/mufon_news/command_line_interface.rb"
+require_relative "../lib/command_line_interface.rb"
 
 
 CommandLineInterface.new
