+++ lib/mufon_news.rb
@@ -1,5 +1,4 @@
 require "mufon_news/version"
 
 module MufonNews
-  CommandLineInterface.new
 end
