+++ lib/sighting_reports.rb
