+++ mufon_news.gemspec
@@ -7,12 +7,13 @@ Gem::Specification.new do |spec|
   spec.name          = "mufon_news"
   spec.version       = MufonNews::VERSION
   spec.authors       = ["Trevor Seitz"]
-  spec.email         = ["trevinla@gmail.com\n"]
+  spec.email         = ["trevinla@gmail.com"]
 
-  spec.summary       = A CLI Web Scraper for the mufon.com website
-  spec.description   = This CLI offers the options to read the most recent news stories about UFOs or MUFON. 
-                        It also offers you the chance to see the most recent UFO sighting reports
-  spec.homepage      = https://github.com/TrevorSeitz/mufon_news
+  spec.summary       = "A CLI Web Scraper for the mufon.com website"
+  spec.description   = "This CLI offers the options to read the most recent news stories about UFOs or MUFON. 
+                        It also offers you the chance to see the most recent UFO sighting reports"
+  spec.homepage      = "https://github.com/TrevorSeitz/mufon_news"
+  
   spec.license       = "MIT"
 
   # Prevent pushing this gem to RubyGems.org. To allow pushes either set the 'allowed_push_host'
@@ -28,11 +29,9 @@ Gem::Specification.new do |spec|
     f.match(%r{^(test|spec|features)/})
   end
   spec.bindir        = "bin"
-  spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
+  spec.executables   = spec.files.grep(/^bin\//) { "run_mufon_scraper "}
   spec.require_paths = ["lib"]
 
   spec.add_development_dependency "bundler", "~> 1.16"
-  spec.add_development_dependency "rake", "~> 10.0"
-  spec.add_development_dependency "rspec", "~> 3.0"
-  spec.add_dependency "nokogiri", "~> 1.8.2"
+  spec.add_dependency "nokogiri", "~> 1.8"
 end
